選択削除時のバグ解消
