+++ resources/js/Components/PaginateByFront.jsx
@@ -3,7 +3,7 @@ export default function Pagination({ page, limit, clickPage, footer, }) {
     return (
         <div className=
             { footer ?
-            "bg-white px-4 py-3 flex items-center justify-between border-t border-gray-200 sm:px-6"
+            "bg-white px-4 py-3 flex items-center  justify-between border-t border-gray-200 sm:px-6"
             :
             ''   
             }
